Connect Rolodex Angular to Sprockets
Have Sprockets load up Rolodex Angular. The way this is loaded is intended to allow for a future Rolodex jQuery version. Update README with some basic info as to what this collection of modules is and how to use it.

Sample usage of how loading Rolodex Angular vs jQuery would work

```coffeescript
#= require 'rolodex_angular'
#= require 'rolodex_jquery'
```